break up GrpcWindmillServer.java to several files
