+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/windmill/grpcclient/GrpcGetWorkStream.java
@@ -1,0 +1,252 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.beam.runners.dataflow.worker.windmill.grpcclient;
+
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicLong;
+import javax.annotation.Nullable;
+import org.apache.beam.runners.dataflow.worker.WindmillTimeUtils;
+import org.apache.beam.runners.dataflow.worker.windmill.AbstractWindmillStream;
+import org.apache.beam.runners.dataflow.worker.windmill.CloudWindmillServiceV1Alpha1Grpc;
+import org.apache.beam.runners.dataflow.worker.windmill.StreamObserverFactory;
+import org.apache.beam.runners.dataflow.worker.windmill.Windmill;
+import org.apache.beam.runners.dataflow.worker.windmill.Windmill.GetWorkRequest;
+import org.apache.beam.runners.dataflow.worker.windmill.Windmill.LatencyAttribution;
+import org.apache.beam.runners.dataflow.worker.windmill.Windmill.StreamingGetWorkRequest;
+import org.apache.beam.runners.dataflow.worker.windmill.Windmill.StreamingGetWorkRequestExtension;
+import org.apache.beam.runners.dataflow.worker.windmill.Windmill.StreamingGetWorkResponseChunk;
+import org.apache.beam.runners.dataflow.worker.windmill.WindmillServerStub.GetWorkStream;
+import org.apache.beam.runners.dataflow.worker.windmill.WindmillServerStub.WorkItemReceiver;
+import org.apache.beam.sdk.util.BackOff;
+import org.apache.beam.vendor.grpc.v1p54p0.com.google.protobuf.ByteString;
+import org.joda.time.Instant;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+final class GrpcGetWorkStream
+    extends AbstractWindmillStream<StreamingGetWorkRequest, StreamingGetWorkResponseChunk>
+    implements GetWorkStream {
+
+  private static final Logger LOG = LoggerFactory.getLogger(GrpcGetWorkStream.class);
+
+  private final GetWorkRequest request;
+  private final WorkItemReceiver receiver;
+  private final ThrottleTimer getWorkThrottleTimer;
+  private final Map<Long, GrpcGetWorkStream.WorkItemBuffer> buffers;
+  private final AtomicLong inflightMessages;
+  private final AtomicLong inflightBytes;
+
+  private GrpcGetWorkStream(
+      CloudWindmillServiceV1Alpha1Grpc.CloudWindmillServiceV1Alpha1Stub stub,
+      GetWorkRequest request,
+      BackOff backoff,
+      StreamObserverFactory streamObserverFactory,
+      Set<AbstractWindmillStream<?, ?>> streamRegistry,
+      int logEveryNStreamFailures,
+      ThrottleTimer getWorkThrottleTimer,
+      WorkItemReceiver receiver) {
+    super(
+        responseObserver ->
+            stub.withDeadlineAfter(
+                    AbstractWindmillStream.DEFAULT_STREAM_RPC_DEADLINE_SECONDS, TimeUnit.SECONDS)
+                .getWorkStream(responseObserver),
+        backoff,
+        streamObserverFactory,
+        streamRegistry,
+        logEveryNStreamFailures);
+    this.request = request;
+    this.getWorkThrottleTimer = getWorkThrottleTimer;
+    this.receiver = receiver;
+    this.buffers = new ConcurrentHashMap<>();
+    this.inflightMessages = new AtomicLong();
+    this.inflightBytes = new AtomicLong();
+  }
+
+  static GrpcGetWorkStream create(
+      CloudWindmillServiceV1Alpha1Grpc.CloudWindmillServiceV1Alpha1Stub stub,
+      GetWorkRequest request,
+      BackOff backoff,
+      StreamObserverFactory streamObserverFactory,
+      Set<AbstractWindmillStream<?, ?>> streamRegistry,
+      int logEveryNStreamFailures,
+      ThrottleTimer getWorkThrottleTimer,
+      WorkItemReceiver receiver) {
+    GrpcGetWorkStream getWorkStream =
+        new GrpcGetWorkStream(
+            stub,
+            request,
+            backoff,
+            streamObserverFactory,
+            streamRegistry,
+            logEveryNStreamFailures,
+            getWorkThrottleTimer,
+            receiver);
+    getWorkStream.startStream();
+    return getWorkStream;
+  }
+
+  private void sendRequestExtension(long moreItems, long moreBytes) {
+    final StreamingGetWorkRequest extension =
+        StreamingGetWorkRequest.newBuilder()
+            .setRequestExtension(
+                StreamingGetWorkRequestExtension.newBuilder()
+                    .setMaxItems(moreItems)
+                    .setMaxBytes(moreBytes))
+            .build();
+
+    executor()
+        .execute(
+            () -> {
+              try {
+                send(extension);
+              } catch (IllegalStateException e) {
+                // Stream was closed.
+              }
+            });
+  }
+
+  @Override
+  protected synchronized void onNewStream() {
+    buffers.clear();
+    inflightMessages.set(request.getMaxItems());
+    inflightBytes.set(request.getMaxBytes());
+    send(StreamingGetWorkRequest.newBuilder().setRequest(request).build());
+  }
+
+  @Override
+  protected boolean hasPendingRequests() {
+    return false;
+  }
+
+  @Override
+  public void appendSpecificHtml(PrintWriter writer) {
+    // Number of buffers is same as distinct workers that sent work on this stream.
+    writer.format(
+        "GetWorkStream: %d buffers, %d inflight messages allowed, %d inflight bytes allowed",
+        buffers.size(), inflightMessages.intValue(), inflightBytes.intValue());
+  }
+
+  @Override
+  public void sendHealthCheck() {
+    send(
+        StreamingGetWorkRequest.newBuilder()
+            .setRequestExtension(
+                StreamingGetWorkRequestExtension.newBuilder().setMaxItems(0).setMaxBytes(0).build())
+            .build());
+  }
+
+  @Override
+  protected void onResponse(StreamingGetWorkResponseChunk chunk) {
+    getWorkThrottleTimer.stop();
+
+    GrpcGetWorkStream.WorkItemBuffer buffer =
+        buffers.computeIfAbsent(
+            chunk.getStreamId(), unused -> new GrpcGetWorkStream.WorkItemBuffer());
+    buffer.append(chunk);
+
+    if (chunk.getRemainingBytesForWorkItem() == 0) {
+      long size = buffer.bufferedSize();
+      buffer.runAndReset();
+
+      // Record the fact that there are now fewer outstanding messages and bytes on the stream.
+      long numInflight = inflightMessages.decrementAndGet();
+      long bytesInflight = inflightBytes.addAndGet(-size);
+
+      // If the outstanding items or bytes limit has gotten too low, top both off with a
+      // GetWorkExtension.  The goal is to keep the limits relatively close to their maximum
+      // values without sending too many extension requests.
+      if (numInflight < request.getMaxItems() / 2 || bytesInflight < request.getMaxBytes() / 2) {
+        long moreItems = request.getMaxItems() - numInflight;
+        long moreBytes = request.getMaxBytes() - bytesInflight;
+        inflightMessages.getAndAdd(moreItems);
+        inflightBytes.getAndAdd(moreBytes);
+        sendRequestExtension(moreItems, moreBytes);
+      }
+    }
+  }
+
+  @Override
+  protected void startThrottleTimer() {
+    getWorkThrottleTimer.start();
+  }
+
+  private class WorkItemBuffer {
+    private final GetWorkTimingInfosTracker workTimingInfosTracker;
+    private String computation;
+    @Nullable private Instant inputDataWatermark;
+    @Nullable private Instant synchronizedProcessingTime;
+    private ByteString data;
+    private long bufferedSize;
+
+    @SuppressWarnings("initialization.fields.uninitialized")
+    WorkItemBuffer() {
+      workTimingInfosTracker = new GetWorkTimingInfosTracker(System::currentTimeMillis);
+      data = ByteString.EMPTY;
+      bufferedSize = 0;
+    }
+
+    @SuppressWarnings("NullableProblems")
+    private void setMetadata(Windmill.ComputationWorkItemMetadata metadata) {
+      this.computation = metadata.getComputationId();
+      this.inputDataWatermark =
+          WindmillTimeUtils.windmillToHarnessWatermark(metadata.getInputDataWatermark());
+      this.synchronizedProcessingTime =
+          WindmillTimeUtils.windmillToHarnessWatermark(
+              metadata.getDependentRealtimeInputWatermark());
+    }
+
+    private void append(StreamingGetWorkResponseChunk chunk) {
+      if (chunk.hasComputationMetadata()) {
+        setMetadata(chunk.getComputationMetadata());
+      }
+
+      this.data = data.concat(chunk.getSerializedWorkItem());
+      this.bufferedSize += chunk.getSerializedWorkItem().size();
+      workTimingInfosTracker.addTimingInfo(chunk.getPerWorkItemTimingInfosList());
+    }
+
+    private long bufferedSize() {
+      return bufferedSize;
+    }
+
+    private void runAndReset() {
+      try {
+        Windmill.WorkItem workItem = Windmill.WorkItem.parseFrom(data.newInput());
+        List<LatencyAttribution> getWorkStreamLatencies =
+            workTimingInfosTracker.getLatencyAttributions();
+        receiver.receiveWork(
+            computation,
+            inputDataWatermark,
+            synchronizedProcessingTime,
+            workItem,
+            getWorkStreamLatencies);
+      } catch (IOException e) {
+        LOG.error("Failed to parse work item from stream: ", e);
+      }
+      workTimingInfosTracker.reset();
+      data = ByteString.EMPTY;
+      bufferedSize = 0;
+    }
+  }
+}
